Adding error messsages

+++ src/components/TaskCard/TaskCard.tsx
@@ -1,10 +1,8 @@
-
 import { taskService } from "@/service/taskService";
 import Link from "next/link";
 
 interface TaskCardProps {
-  
-  id:  string ;
+  id: string;
   title: string;
   time?: string;
   description: string;
@@ -12,30 +10,28 @@ interface TaskCardProps {
 }
 
 const TaskCard: React.FC<TaskCardProps> = ({
-  
   id,
   title,
   time,
   description,
   onDelete,
 }) => {
-
-
   const deleteTask = async (id: string) => {
     try {
-      const response = await taskService.deleteTask(id);
-      alert(response)
-      onDelete(id)
-    } catch (error) {
-      alert(error)
+      if (window.confirm("Are you sure you want to delete this task?")) {
+        const response = await taskService.deleteTask(id);
+        alert(response);
+        onDelete(id);
+      }
+
+      
+    } catch (error: any) {
+      alert(error.message);
     }
-  }
-
-
-
+  };
 
   return (
-    <div  className="bg-gray-200 p-4 flex flex-col justify-between">
+    <div className="bg-gray-200 p-4 flex flex-col justify-between">
       <section className="text-red mb-4">
         <div className="flex justify-between items-center">
           <h3 className="text-2xl">{title}</h3>
@@ -66,4 +62,4 @@ const TaskCard: React.FC<TaskCardProps> = ({
   );
 };
 
-export default TaskCard
+export default TaskCard;
